+++ vec_math_bonus.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   vec3_math_bonus.c                                  :+:      :+:    :+:   */
+/*   vec_math_bonus.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: hermarti <hermarti@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/09 15:41:33 by hermarti          #+#    #+#             */
-/*   Updated: 2025/09/09 16:03:19 by hermarti         ###   ########.fr       */
+/*   Updated: 2025/09/10 02:42:18 by hermarti         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ t_vec3	ft_vec3_add(t_vec3 a, t_vec3 b)
 	return (res);
 }
 
-t_vec3	ft_vec3_scale(t_vec3 v, double s)
+t_vec3	ft_vec3_scale(t_vec3 v, float s)
 {
 	t_vec3	res;
 
@@ -33,7 +33,7 @@ t_vec3	ft_vec3_scale(t_vec3 v, double s)
 	return (res);
 }
 
-double	ft_vec3_magnitude(t_vec3 v)
+float	ft_vec3_magnitude(t_vec3 v)
 {
 	return (sqrt((ft_vec3_dot(v, v))));
 }
@@ -41,13 +41,13 @@ double	ft_vec3_magnitude(t_vec3 v)
 t_vec3	ft_vec3_normalize(t_vec3 v)
 {
 	t_vec3	res;
-	double	mag;
+	float	mag;
 
 	mag = ft_vec3_magnitude(v);
-	res.x = 0.0;
-	res.y = 0.0;
-	res.z = 0.0;
-	if (mag == 0.0)
+	res.x = 0.0f;
+	res.y = 0.0f;
+	res.z = 0.0f;
+	if (mag == 0.0f)
 		return (res);
 	res.x = v.x / mag;
 	res.y = v.y / mag;
@@ -55,7 +55,7 @@ t_vec3	ft_vec3_normalize(t_vec3 v)
 	return (res);
 }
 
-double	ft_vec3_dot(t_vec3 a, t_vec3 b)
+float	ft_vec3_dot(t_vec3 a, t_vec3 b)
 {
 	return ((a.x * b.x) + (a.y * b.y) + (a.z * b.z));
 }
